Refactor and export types across components for improved type management and code clarity
