+++ components/admin/AdminSidebar.types.ts
@@ -1,0 +1,4 @@
+export type AdminSidebarProps = {
+  adminName: string | null;
+  adminEmail: string;
+};
