+++ components/insights/WeeklyUsageChart.types.ts
@@ -1,0 +1,9 @@
+export type WeeklyUsagePoint = {
+  day: string;
+  thisWk: number;
+  lastWk: number;
+};
+
+export type WeeklyUsageChartProps = {
+  data: WeeklyUsagePoint[];
+};
