+++ components/realtime/RealtimeRefreshBridge.types.ts
@@ -1,0 +1,5 @@
+export type RealtimeRefreshBridgeProps = {
+  deviceKeys: string[];
+  throttleMs?: number;
+  pollMs?: number;
+};
