+++ components/ui/HomeBudgetEditor.types.ts
@@ -1,0 +1,3 @@
+export type HomeBudgetEditorProps = {
+  initialBudget: number;
+};
