+++ components/ui/LogoutButton.types.ts
@@ -1,0 +1,4 @@
+export type LogoutButtonProps = {
+  className?: string;
+  label?: string;
+};
