+++ components/ui/MobileViewport.types.ts
@@ -1,0 +1,5 @@
+import type { ReactNode } from "react";
+
+export type MobileViewportProps = {
+  children: ReactNode;
+};
