+++ components/ui/RelayToggle.types.ts
@@ -1,0 +1,5 @@
+export type RelayToggleProps = {
+  deviceId: string;
+  initialRelayState: boolean;
+  variant?: "compact" | "full";
+};
